Auto-fuzz: Extract Oss-fuzz-project class
