+++ tools/auto-fuzz/manager.py
@@ -40,6 +40,7 @@ import oss_fuzz_manager
 import fuzz_driver_generation_python
 import fuzz_driver_generation_jvm
 import post_process
+from classes import OSS_FUZZ_PROJECT
 
 from io import BytesIO
 from typing import List, Any
@@ -66,116 +67,6 @@ if not os.path.isfile(FUZZ_INTRO_MAIN["jvm"]):
 
 if not os.path.isdir(OSS_FUZZ_BASE):
     raise Exception("Could not find OSS-Fuzz directory")
-
-
-class OSS_FUZZ_PROJECT:
-    """Abstraction of OSS-Fuzz project.
-
-    Provides helper methods for easily managing files and folders and
-    operations on a given OSS-Fuzz project.
-    """
-
-    def __init__(self, project_folder, github_url, language, benchmark=False):
-        self.project_folder = project_folder
-        self.github_url = github_url
-        self.language = language
-        self.benchmark = benchmark
-
-    @property
-    def build_script(self):
-        return self.project_folder + "/build.sh"
-
-    @property
-    def dockerfile(self):
-        return self.project_folder + "/Dockerfile"
-
-    @property
-    def project_yaml(self):
-        return self.project_folder + "/project.yaml"
-
-    @property
-    def base_fuzzer(self):
-        if self.language == "python":
-            return self.project_folder + "/fuzz_1.py"
-        elif self.language == "jvm":
-            return self.project_folder + "/Fuzz.java"
-        else:
-            # Temporary fail safe logic
-            return self.project_folder + "/fuzz_1.py"
-
-    @property
-    def oss_fuzz_project_name(self):
-        return os.path.basename(self.project_folder)
-
-    @property
-    def oss_fuzz_fuzzer_namer(self):
-        if self.language == "python":
-            return os.path.basename(self.base_fuzzer).replace(".py", "")
-        elif self.language == "jvm":
-            return os.path.basename(self.base_fuzzer).replace(".java", "")
-        else:
-            # Temporary fail safe logic
-            return os.path.basename(self.base_fuzzer).replace(".py", "")
-
-    @property
-    def project_name(self):
-        if self.benchmark:
-            return "%s-%s" % (self.language, self.github_url)
-        else:
-            # Simplify url by cutting https out, then assume what we have left is:
-            # HTTP Type
-            # github.com/{user}/{proj_name}
-            # or
-            # SSH Type
-            # git@github.com:{user}/{proj_name}
-            if self.github_url.startswith("https://"):
-                return self.github_url.replace("https://", "").split("/")[2]
-            elif self.github_url.startswith("http://"):
-                return self.github_url.replace("http://", "").split("/")[2]
-            else:
-                return self.github_url.split("/")[1]
-
-    def write_basefiles(self, project_build_type=None):
-        with open(self.build_script, "w") as builder_file:
-            builder_file.write(
-                base_files.gen_builder_1(self.language, project_build_type))
-
-        with open(self.base_fuzzer, "w") as fuzzer_file:
-            fuzzer_file.write(
-                base_files.gen_base_fuzzer(self.language, project_build_type))
-
-        with open(self.project_yaml, "w") as yaml_file:
-            yaml_file.write(
-                base_files.gen_project_yaml(self.github_url, self.language,
-                                            project_build_type))
-
-        with open(self.dockerfile, "w") as docker_file:
-            docker_file.write(
-                base_files.gen_dockerfile(
-                    self.github_url,
-                    self.project_name,
-                    self.language,
-                    project_build_type=project_build_type))
-
-    def change_jvm_dockerfile(self,
-                              jdk_version,
-                              project_build_type,
-                              build_project=True):
-        with open(self.dockerfile, "w") as docker_file:
-            docker_file.write(
-                base_files.gen_dockerfile(
-                    self.github_url,
-                    self.project_name,
-                    self.language,
-                    jdk_version,
-                    build_project,
-                    project_build_type=project_build_type))
-
-    def change_build_script(self, project_build_type, build_project=True):
-        with open(self.build_script, "w") as builder_file:
-            builder_file.write(
-                base_files.gen_builder_1(self.language, project_build_type,
-                                         build_project))
 
 
 def get_next_project_folder(base_dir):
